+++ sbk-api/src/main/java/io/sbk/perl/impl/ArrayLatencyRecorder.java
@@ -87,7 +87,7 @@ public class ArrayLatencyRecorder extends LatencyWindow {
      * @param latency latency value in milliseconds.
      */
     @Override
-    public void record(long startTime, int bytes, int events, long latency) {
+    public void record(long startTime, long bytes, long events, long latency) {
         if (record(bytes, events, latency)) {
             final int index = (int) (latency - this.lowLatency);
             this.minIndex = Math.min(this.minIndex, index);
